algo_classifier/requirements.txt
fastapi
python-multipart
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
PyJWT
passlib[bcrypt]
tenacity
google-generativeai
email-validator
bcrypt==3.2.2